+++ .pi/extensions/pi-doctor.ts
@@ -1,0 +1,1 @@
+export { default } from "../../extensions/pi-doctor.ts";